search

+++ src/zobrist.rs
@@ -3,16 +3,16 @@ use lazy_static::lazy_static;
 use crate::board::{Board, Color, Piece, Square};
 use crate::utils::pseudo_random;
 
-fn gen_pk() -> [u32; 13 * 128] {
-    let mut temp = [0u32; 13 * 128];
+fn gen_pk() -> [u64; 13 * 128] {
+    let mut temp = [0; 13 * 128];
     for i in 0..13 * 128 {
         temp[i] = pseudo_random();
     }
     temp
 }
 
-fn gen_castle() -> [u32; 16] {
-    let mut temp = [0u32; 16];
+fn gen_castle() -> [u64; 16] {
+    let mut temp = [0; 16];
     for i in 0..16 {
         temp[i] = pseudo_random();
     }
@@ -20,14 +20,14 @@ fn gen_castle() -> [u32; 16] {
 }
 
 lazy_static! {
-    pub static ref PIECE_KEYS: [u32; 13 * 128] = gen_pk();
-    pub static ref CASTLE_KEYS: [u32; 16] = gen_castle();
-    pub static ref SIDE_KEY: u32 = pseudo_random();
+    pub static ref PIECE_KEYS: [u64; 13 * 128] = gen_pk();
+    pub static ref CASTLE_KEYS: [u64; 16] = gen_castle();
+    pub static ref SIDE_KEY: u64 = pseudo_random();
 }
 
 impl Board {
-    pub fn gen_hash(&self) -> u32 {
-        let mut res = 0;
+    pub fn gen_hash(&self) -> u64 {
+        let mut res = 0u64;
 
         for sq in 0..128 {
             if sq & 0x88 == 0 {
